python-sdk/pyproject.toml
[project]
name = "fantasy-climbing-sdk"
version = "0.1.0"
description = "Python SDK for Fantasy Climbing League - IFSC data sync and management"
authors = [
    {name = "FelixAbrahamsson", email = "felixabrahamsson1@gmail.com"}
]
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.0",
    "supabase>=2.27.0",
    "python-dotenv>=1.2.0",
    "pydantic>=2.0.0",
]

[project.scripts]
fcl-sync = "fantasy_climbing.cli:main"

[tool.poetry]
name = "fantasy-climbing-sdk"
version = "0.1.0"
description = "Python SDK for Fantasy Climbing League - IFSC data sync and management"
authors = ["FelixAbrahamsson <felixabrahamsson1@gmail.com>"]
readme = "README.md"
packages = [
    { include = "ifsc_sdk" },
    { include = "fantasy_climbing" }
]

[tool.poetry.dependencies]
python = ">=3.10"
httpx = ">=0.28.0"
supabase = ">=2.27.0"
python-dotenv = ">=1.2.0"
pydantic = ">=2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0.0"
pytest-asyncio = ">=0.23.0"

[tool.poetry.scripts]
fcl-sync = "fantasy_climbing.cli:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
